+++ js/util.js
@@ -1,0 +1,24 @@
+//Функция, возвращающая случайное число в заданном диапазоне (включительно)
+const getRandomInteger = function(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1) + min);
+};
+
+//Функция, возвращающая уникальные значения из заданного диапазона
+const createUniqueRandomInteger = function (min, max) {
+  const previousValues = [];
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+export {getRandomInteger, createUniqueRandomInteger};
